model_factory: Decouple
